fix ui spc and smc

+++ database/migrations/2024_05_17_160827_create_orderlktis_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     {
         Schema::create('orderlktis', function (Blueprint $table) {
             $table->id();
-            $table->integer('order');
+            $table->bigInteger('order');
             $table->string('kompetisi');
             $table->string('nama');
             $table->string('email');
